refactor(Repeater): Split ItemsRepeater per winui-port layout
Reorganize ItemsRepeater to match the WinUI source layout (commit 5f9e85113):
- ItemsRepeater.cs: class declaration only, reduced to 16 lines
- ItemsRepeater.h.mux.cs: constants, inline properties, fields, events in WinUI h order
- ItemsRepeater.mux.cs: method bodies in cpp source order (constructor, OnLoaded/OnUnloaded, OnItemsSourceViewChanged, OnLayoutChanged, GetElementImpl, ...)
- ItemsRepeater.uno.cs: Uno-only additions (SerialDisposable revokers for StackLayout leak prevention, UIElementCollection/IPanel contract, UnoBeforeElementPrepared diagnostic event, OnLoadedUno/OnUnloadedUno for native re-subscription)

Add GetEffectiveLayout/GetDefaultLayout with thread-local default StackLayout fallback. Port MustClearDataContext wiring. Native-only forced relayout scoped via UNO_HAS_ENHANCED_LIFECYCLE (Skia runs the WinUI path).
